saqlain2022skipq/Training_Sprint_7/Design_Problem_2/requirements.txt
aws-cdk-lib==2.69.0
constructs>=10.0.0,<11.0.0
cdk_ecr_deployment
